nf: extract BackendUtils
Functions will be used by other classes
